chore(tauri/ui): move vault / order service calls into same file
